Backend/requirements.txt
asgiref
Django
djangorestframework
djangorestframework_simplejwt
filelock
fsspec
gunicorn
Jinja2
MarkupSafe
mpmath
networkx
packaging
pillow
psycopg2-binary
PyJWT
python-dotenv
setuptools
sqlparse
sympy
typing_extensions
